+++ examples/asynchronous_telebot/middleware/flooding_middleware.py
@@ -1,6 +1,6 @@
 # Just a little example of middleware handlers
 
-from telebot.asyncio_handler_backends import BaseMiddleware
+from telebot.asyncio_middlewares import BaseMiddleware
 from telebot.async_telebot import AsyncTeleBot
 from telebot.async_telebot import CancelUpdate
 bot = AsyncTeleBot('TOKEN')
