+++ examples/asynchronous_telebot/middleware/i18n.py
@@ -7,6 +7,7 @@
 # But this example just to show the work of middlewares.
 
 import telebot
+import telebot.asyncio_middlewares
 from telebot.async_telebot import AsyncTeleBot
 from telebot import asyncio_handler_backends
 import logging
@@ -27,7 +28,7 @@ TRANSLATIONS = {
 bot = AsyncTeleBot('TOKEN')
 
 
-class LanguageMiddleware(asyncio_handler_backends.BaseMiddleware):
+class LanguageMiddleware(telebot.asyncio_middlewares.BaseMiddleware):
     def __init__(self):
         self.update_types = ['message'] # Update types that will be handled by this middleware.
     async def pre_process(self, message, data):
