+++ telebot/asyncio_handler_backends.py
@@ -1,18 +1,3 @@
-class BaseMiddleware:
-    """
-    Base class for middleware.
-
-    Your middlewares should be inherited from this class.
-    """
-    def __init__(self):
-        pass
-
-    async def pre_process(self, message, data):
-        raise NotImplementedError
-    async def post_process(self, message, data, exception):
-        raise NotImplementedError
-
-
 class State:
     def __init__(self) -> None:
         self.name = None
